refactor(spec): make State fork-agnostic for amsterdam
Move the `State` class and all state helper functions from
`ethereum/forks/amsterdam/state.py` into the shared `ethereum/state.py`,
making the state implementation fork-agnostic for amsterdam.

To break the circular import that would arise between `state.py` and
`trie.py`, introduce `ethereum/types.py` as a thin module holding the
base account types (`Account`, `Address`, `Root`, `EMPTY_CODE_HASH`,
`EMPTY_ACCOUNT`). `trie.py` now imports from `types.py`, allowing
`state.py` to import from both `types.py` and `trie.py` without a cycle.

`ethereum/state.py` re-exports the base types so that the ~260 existing
`from ethereum.state import ...` call sites across all forks require no
changes.

`fork_loader.py` and `t8n_types.py` are updated to fall back to the
shared `ethereum.state` module when the fork-specific `state` module no
longer exists.
