+++ src/ethereum_spec_tools/evm_tools/loaders/fork_loader.py
@@ -307,37 +307,72 @@ class ForkLoad:
     @property
     def State(self) -> Any:
         """State class of the fork."""
-        return self._module("state").State
+        try:
+            return self._module("state").State
+        except ModuleNotFoundError:
+            from ethereum.state import State
+
+            return State
 
     @property
     def set_account(self) -> Any:
         """set_account function of the fork."""
-        return self._module("state").set_account
+        try:
+            return self._module("state").set_account
+        except ModuleNotFoundError:
+            from ethereum.state import set_account
+
+            return set_account
 
     @property
     def store_code(self) -> Any:
         """store_code function of the fork."""
-        return getattr(self._module("state"), "store_code", None)
+        try:
+            return getattr(self._module("state"), "store_code", None)
+        except ModuleNotFoundError:
+            from ethereum.state import store_code
+
+            return store_code
 
     @property
     def set_storage(self) -> Any:
         """set_storage function of the fork."""
-        return self._module("state").set_storage
+        try:
+            return self._module("state").set_storage
+        except ModuleNotFoundError:
+            from ethereum.state import set_storage
+
+            return set_storage
 
     @property
     def state_root(self) -> Any:
         """state_root function of the fork."""
-        return self._module("state").state_root
+        try:
+            return self._module("state").state_root
+        except ModuleNotFoundError:
+            from ethereum.state import state_root
+
+            return state_root
 
     @property
     def close_state(self) -> Any:
         """close_state function of the fork."""
-        return self._module("state").close_state
+        try:
+            return self._module("state").close_state
+        except ModuleNotFoundError:
+            from ethereum.state import close_state
+
+            return close_state
 
     @property
     def create_ether(self) -> Any:
         """create_ether function of the fork."""
-        return self._module("state").create_ether
+        try:
+            return self._module("state").create_ether
+        except ModuleNotFoundError:
+            import ethereum.state
+
+            return getattr(ethereum.state, "create_ether", None)
 
     @property
     def root(self) -> Any:
